Implemented JSON Export:
Provide a button that allows users to export their quotes to a JSON file. Use Blob and URL.createObjectURL to create a download link.
Implemented JSON Import:

Provide a file input to allow users to upload a JSON file containing quotes. Read the file and update the quotes array and local storage accordingly.